fix: information text not rendering html
Due to starting the information text in the config.js with a line break, Companion was not rendering the html and was showing it as text.  By replacing the line break with an html line break, Companion will render the html.  This might also be a Companion v4.0 bug causing the rendering issue.